# Conflicts:
#	src/main/java/com/team1/efep/service_implementors/BuyerServiceImpl.java

+++ src/main/java/com/team1/efep/service_implementors/BuyerServiceImpl.java
@@ -58,7 +58,7 @@ public class BuyerServiceImpl implements BuyerService {
     public String viewWishlist(HttpSession session, Model model, RedirectAttributes redirectAttributes) {
         Map<String, String> error = new HashMap<>();
         Account account = Role.getCurrentLoggedAccount(session);
-        if (account == null) {
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
             redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(error, "You are not logged in"));
             return "redirect:/login";
         }
@@ -136,7 +136,7 @@ public class BuyerServiceImpl implements BuyerService {
     public String addToWishlist(AddToWishlistRequest request, HttpServletRequest httpServletRequest, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
         Map<String, String> error = new HashMap<>();
         Account account = Role.getCurrentLoggedAccount(session);
-        if (account == null) {
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
             redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(error, "You are not logged in"));
             return "redirect:/login";
         }
@@ -204,7 +204,7 @@ public class BuyerServiceImpl implements BuyerService {
     @Override
     public String updateWishlist(UpdateWishlistRequest request, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
         Account account = Role.getCurrentLoggedAccount(session);
-        if (account == null) {
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
             return "redirect:/login";
         }
         request.setAccountId(account.getId());
@@ -212,11 +212,11 @@ public class BuyerServiceImpl implements BuyerService {
         if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, UpdateWishlistResponse.class)) {
             session.setAttribute("acc", accountRepo.findById(request.getAccountId()).orElse(null));
             redirectAttributes.addFlashAttribute("msg", (UpdateWishlistResponse) output);
-            AllPage.allConfig(model, this);
+            AllPage.allConfig(model, this, session);
             return viewWishlist(session, model, redirectAttributes);
         }
         redirectAttributes.addFlashAttribute("error", (Map<String, String>) output);
-        AllPage.allConfig(model, this);
+        AllPage.allConfig(model, this, session);
         return viewWishlist(session, model, redirectAttributes);
     }
 
@@ -254,6 +254,11 @@ public class BuyerServiceImpl implements BuyerService {
 
     @Override
     public String deleteWishlist(DeleteWishlistRequest request, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
+            redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(new HashMap<>(), "You are not logged in"));
+            return "redirect:/login";
+        }
         Object output = deleteWishlistLogic(request);
         if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, DeleteWishlistResponse.class)) {
             session.setAttribute("acc", accountRepo.findById(request.getAccountId()).orElse(null));
@@ -293,7 +298,7 @@ public class BuyerServiceImpl implements BuyerService {
     public String deleteWishlistItem(DeleteWishlistItemRequest request, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
         Map<String, String> error = new HashMap<>();
         Account account = Role.getCurrentLoggedAccount(session);
-        if (account == null) {
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
             redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(error, "You are not logged in"));
             return "redirect:/login";
         }
@@ -334,6 +339,11 @@ public class BuyerServiceImpl implements BuyerService {
     //-----------------------------------------------FORGOT PASSWORD------------------------------------------------------------//
     @Override
     public String sendEmail(ForgotPasswordRequest request, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
+            redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(new HashMap<>(), "You are not logged in"));
+            return "redirect:/login";
+        }
         Object output = sendEmailLogic(request);
         if (!OutputCheckerUtil.checkIfThisIsAResponseObject(output, ForgotPasswordResponse.class)) {
             redirectAttributes.addFlashAttribute("error",  output);
@@ -412,6 +422,11 @@ public class BuyerServiceImpl implements BuyerService {
 
     @Override
     public String renewPass(RenewPasswordRequest request, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
+            redirectAttributes.addFlashAttribute(MapConfig.buildMapKey(new HashMap<>(), "You are not logged in"));
+            return "redirect:/login";
+        }
         request.setEmail(session.getAttribute("mail").toString());
         Object output = renewPassLogic(request);
         if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, RenewPasswordResponse.class)) {
@@ -444,6 +459,14 @@ public class BuyerServiceImpl implements BuyerService {
     //-------------------------------------------VIEW FLOWER LIST---------------------------------------//
     @Override
     public String viewFlowerList(HttpSession session, Model model) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account != null && Role.checkIfThisAccountIsSeller(account)) {
+            model.addAttribute("error", ViewFlowerListResponse.builder()
+                    .status("400")
+                    .message("No permission")
+                    .build());
+            return "redirect:/login";
+        }
         ViewFlowerListResponse output = viewFlowerListLogic();
         model.addAttribute("msg", output);
         return "category";
@@ -695,7 +718,15 @@ public class BuyerServiceImpl implements BuyerService {
     //--------------------------------------SEARCH FLOWER------------------------------------------//
 
     @Override
-    public String searchFlower(SearchFlowerRequest request, Model model) {
+    public String searchFlower(SearchFlowerRequest request, Model model, HttpSession session) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account != null && Role.checkIfThisAccountIsSeller(account)) {
+            model.addAttribute("error", SearchFlowerResponse.builder()
+                    .status("400")
+                    .message("No permission")
+                    .build());
+            return "redirect:/login";
+        }
         model.addAttribute("msg", searchFlowerLogic(request));
         return "category";
     }
@@ -735,7 +766,15 @@ public class BuyerServiceImpl implements BuyerService {
     //--------------------------------------------VIEW FLOWER DETAIL---------------------------------------------//
 
     @Override
-    public String viewFlowerDetail(ViewFlowerDetailRequest request, Model model) {
+    public String viewFlowerDetail(ViewFlowerDetailRequest request, Model model, HttpSession session) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account != null && Role.checkIfThisAccountIsSeller(account)) {
+            model.addAttribute("error", ViewFlowerDetailResponse.builder()
+                    .status("400")
+                    .message("No permission")
+                    .build());
+            return "redirect:/login";
+        }
         Object output = viewFlowerDetailLogic(request);
         if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, ViewFlowerDetailResponse.class)) {
             model.addAttribute("msg", (ViewFlowerDetailResponse) output);
@@ -798,7 +837,7 @@ public class BuyerServiceImpl implements BuyerService {
     @Override
     public String viewOrderStatus(HttpSession session, Model model, RedirectAttributes redirectAttributes) {
         Account account = Role.getCurrentLoggedAccount(session);
-        if (account == null) {
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
             redirectAttributes.addFlashAttribute("error", "You are not logged in");
             return "redirect:/login";
         }
@@ -833,6 +872,14 @@ public class BuyerServiceImpl implements BuyerService {
 
     @Override
     public String cancelOrder(CancelOrderRequest request, HttpSession session, Model model, HttpServletRequest httpServletRequest, RedirectAttributes redirectAttributes) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account == null || !Role.checkIfThisAccountIsBuyer(account)) {
+            model.addAttribute("error", CancelOrderResponse.builder()
+                    .status("400")
+                    .message("Please login a buyer account to do this action")
+                    .build());
+            return "redirect:/login";
+        }
         String referer = httpServletRequest.getHeader("Referer");
         Object output = cancelOrderLogic(request);
         if (OutputCheckerUtil.checkIfThisIsAResponseObject(output, CancelOrderResponse.class)) {
@@ -929,7 +976,15 @@ public class BuyerServiceImpl implements BuyerService {
     //--------------------------------VIEW CATEGORY------------------------------------------//
 
     @Override
-    public void viewCategory(Model model) {
+    public void viewCategory(Model model, HttpSession session) {
+        Account account = Role.getCurrentLoggedAccount(session);
+        if (account != null && Role.checkIfThisAccountIsSeller(account)) {
+            model.addAttribute("error", ViewFlowerListResponse.builder()
+                    .status("400")
+                    .message("No permission")
+                    .build());
+            return;
+        }
         model.addAttribute("msg3", viewCategoryLogic());
     }
 
@@ -955,7 +1010,7 @@ public class BuyerServiceImpl implements BuyerService {
     //--------------------------------VN Pay------------------------------------------//
 
     @Override
-    public String createVNPayPaymentLink(VNPayRequest request, Model model, HttpServletRequest httpServletRequest) {
+    public String createVNPayPaymentLink(VNPayRequest request, Model model, HttpServletRequest httpServletRequest, HttpSession session) {
         VNPayResponse vnPayResponse = createVNPayPaymentLinkLogic(request, httpServletRequest);
         model.addAttribute("msg", vnPayResponse);
         return "redirect:" + vnPayResponse.getPaymentURL();
